Modularized Python build script
